backend-django/requirements.txt
Django==2.2.2
djangorestframework
markdown
django-filter
drf-renderer-xlsx
python-keycloak
requests
python-jose
django-cors-headers